Began to Add Map Generation Library.
Previously map.h just defined the data model.
It was up to the various things making the maps to actually put the data together, and keep track of everything.

This was dumb.

So, I've began to abstract the map making out a ton.
Like, a lot.

This, naturally, makes it less efficient.
Really, though, that's not something I care about right now.
I can fix that when it's a problem.

It does, though, make everything a ton more future-proof, and portable, and whatever else.

It also makes it a lot cleaner.
The new testMap.c is infinitely easier to understand, especially to one who doesn't know the crazy model I've built.

Now, asumedly, someone can make their own map generation stuffs without having to navigate my model.

Well, they still might have to a bit.
There's definitely more on the horizon.
I can already see, where I'm going, a couple more functions that need to be defined, and a couple more things to be abstracted out.

We'll see.