Added testScan
This function takes in the input file on stdin and outputs it on stdout.

Mostly used for testing the parsing.

For this I have written the outputMap.h and iniOutput.c because I'm awesome.

In order for this to work without unnecessarily initializing GL, I had to take the glLists thing out of initMap.

That's ok, when I need caching, I should do it more generally anyway.