Implemented Stitch Definition.
This is really rough right now.
Really Rough.

For example, it requires all definition to be at the end because it doesn't support definitions for tiles which aren't defined yet.

I think I'm going to need to do some sort of delayed execution or something.

Anyway, because of the way my files are, that is valid, so it's ok.
It just isn't as flexible as it should be.

Also, the algorithm I've got here is by no means great.
It worked in my couple quick examples, and so that was good enough.
Something else will probably be required.