Make Backs of Tiles Translucent.
This is something I've wanted to do for a while.
Mostly, it allows you to see things that are obscured by your view point.

So, in order to do this I split it up so there are now two functions for drawing tiles.
One for drawing the front, and one for drawing the back.

For now they just invert the drawing and change the alpha.
I'm considering in the future that it would be nice to have a way of changing how walls are rendered.
One would be translucent, one would be missing, one would be half, etc.

Also, it's not perfect yet.
As of now there are some problems with depth order.
Like, some walls are invisible when viewed through another wall.

Work in progress.