Transparent Objects Work Better
So, this is a hack.

I don't do any kind of magic.
I've just made it so I render the world forward, then back.

So, all opaque objects are rendered before transparent ones.

It's inefficient, because I walk the whole tree thing every time.
At this point, though, it's not a big deal.
I just wanted to get past this issue so I could move on the other things.

If I need to fix slow rendering later, this might be the first place I visit.